+++ packages/i18n/CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact i18n module, newest chan
 
 ## [2.0.0-rc.1] - 2018-07-09
 
+### Fixed
+
+- `i18n/Uppercase` to apply the designated `casing` prop format to each child instead of only the first child
+
 ### Removed
 
 - `i18n/I18nDecorator.contextTypes`, replaced by `i18n/I18nDecorator.I18nContextDecorator`